Module lockfree::hazard

Hazard pointer API. Please, note that this API is not implemented in the "traditional" way hazard pointers are implemented. This is because this implemention aims to be lock-free.

Re-exports

pub use std::sync::atomic::Ordering::*;

Structs

HazardPtr
A hazard atomic pointer. It keeps a destructor with itself. The destruction and loading of this pointer uses the hazard API of this module. Because it (possibly later) calls the destructor on drop, be very careful. In general, if the pointer is consumed, you may want to fill the hazard pointer with std::ptr::null_mut(), because HazardPtr checks for nulls before putting the pointer on the queue.

Enums

Ordering
Atomic memory orderings

Functions

critical
Executes the given function as critical code. No deletions of new queues will start during this execution.

later_drop⚠
Adds the given pointer and drop function to the local deletion queue. If there is no critical code executing, the local queue items are deleted. The function is unsafe because pointers must be correctly dropped such as no "use after free" or "double free" happens.

try_delete_local
Tries to delete the local queue items.