Module lockfree::tls

A wait-free per-object Thread Local Storage (TLS).

Structs

IntoIter

An iterator over owned references to entries of TLS.

Iter

An iterator over immutable references to entries of TLS.

IterMut

An iterator over mutable references to entries of TLS.

ThreadId

A cached thread-id. Repeated calls to ThreadLocal’s methods with cached IDs should be faster than reloading the ID everytime.

ThreadLocal

Per Object Thread Local Storage. The stored data is not dropped on thread exit. It is only dropped when the structure itself is dropped. After the thread exited, the data might be reused for other threads. This TLS’s operation are also wait-free.